+++ clients/python-wrapper/lakefs/reference.py
@@ -6,13 +6,10 @@ from __future__ import annotations
 
 from typing import Optional, Generator, Literal
 
-import lakefs_sdk
-
 from lakefs.client import Client, DefaultClient
 from lakefs.exceptions import api_exception_handler
 from lakefs.object import StoredObject
 from lakefs.namedtuple import LenientNamedTuple
-from lakefs.object import Object
 from lakefs.object_manager import ObjectManager
 
 
@@ -88,7 +85,7 @@ class Reference:
                     if max_amount is not None and count >= max_amount:
                         return
 
-    def log(self, max_amount: Optional[int] = None, **kwargs) -> Generator[lakefs_sdk.Commit]:
+    def log(self, max_amount: Optional[int] = None, **kwargs) -> Generator[Commit]:
         """
         Returns a generator of commits starting with this reference id
