+++ clients/python-wrapper/tests/.pylintrc
@@ -13,6 +13,7 @@ disable=
     missing-module-docstring,
     too-few-public-methods,
     pointless-statement,
+    no-member,
     fixme,
 
 [MISCELLANEOUS]
