+++ clients/python-wrapper/tests/integration/conftest.py
@@ -2,23 +2,12 @@ import os
 import re
 import time
 import uuid
-from contextlib import contextmanager
 import pytest
 
 from lakefs import client
 from lakefs.repository import Repository
 
 TEST_STORAGE_NAMESPACE_BASE = os.getenv("STORAGE_NAMESPACE", "").rstrip("/")
-
-
-@contextmanager
-def expect_exception_context(ex, status_code=None):
-    try:
-        yield
-        assert False, f"No exception raised! Expected exception of type {ex.__name__}"
-    except ex as e:
-        if status_code is not None:
-            assert e.status_code == status_code
 
 
 @pytest.fixture(name="test_name", autouse=True)
